move stuff out of earth, clean up
